refactor: resource modal and hooks
