+++ frontend/src/store/index.ts
@@ -38,7 +38,7 @@ interface State {
   searchParams: ISearchParameters;
   treeData: TreeNode;
   selectedNodesIds: string[];
-  currentFolder: Partial<IFolder> | undefined;
+  currentFolder: Partial<IFolder>;
   selectedFolders: IFolder[];
   selectedResources: IResource[];
   folderIds: ID[];
